feat(decoder): внедрение контекстного API для изоляции состояния декодера
- Реализована структура `G723DecoderContext` и функции управления жизненным циклом
- Переведена функция `g723_decode_frame` на использование внешнего контекста
- Настроена сборка в CMake с принудительным форсированием 32-битной архитектуры Win32
- Выделена библиотека кодека `g723_codec` с фильтрацией исходных файлов в CMakeLists.txt
- Устранены предупреждения компилятора при приведении типов в G.711u и G.723.1
- Добавлен тестовый сценарий производительности транскодирования и файл README.md

+++ g711u_coder.cpp
@@ -15,7 +15,7 @@ static unsigned char calculate_pcmu_sample(short sample) {
     for (int mask = 0x4000; (sample & mask) == 0 && exponent > 0; mask >>= 1) {
         exponent--;
     }
-    int mantissa = (sample >> (exponent + 3)) & 0x0F;
+    int mantissa = (static_cast<unsigned short>(sample) >> (exponent + 3)) & 0x0F;
     return ~(sign | (exponent << 4) | mantissa);
 }
